requirements.txt
opencv-python
numpy
pytesseract
mss
pyautogui
pyyaml
loguru

#cd "C:\Users\Spinola\OneDrive\Laboratorio\VSCode\PokeBot-Pro\PokeBot_Pro"
#python run_bot.py